Add link to phabricator to README.md
Since we are not using GitHub's issue functionality for this repository,
we should provide users with an alternative way.

The url in the package.json is to our phabricator instance in general to
allow people to search for bugs before opening a new ticket. That seems to be
the general approach with this field in the industry.